# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Calulator Service
    description: Service for multiplying numbers, a Goa teaser
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /addition/{a}/{b}:
        get:
            tags:
                - calc
            summary: addition calc
            operationId: calc#addition
            parameters:
                - name: a
                  in: path
                  description: Left operand
                  required: true
                  type: integer
                - name: b
                  in: path
                  description: Right operand
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CalcAdditionResponseBody'
                        required:
                            - result
            schemes:
                - http
    /division/{a}/{b}:
        get:
            tags:
                - calc
            summary: division calc
            operationId: calc#division
            parameters:
                - name: a
                  in: path
                  description: Left operand
                  required: true
                  type: integer
                - name: b
                  in: path
                  description: Right operand
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CalcDivisionResponseBody'
                        required:
                            - result
            schemes:
                - http
    /multiply/{a}/{b}:
        get:
            tags:
                - calc
            summary: multiply calc
            operationId: calc#multiply
            parameters:
                - name: a
                  in: path
                  description: Left operand
                  required: true
                  type: integer
                - name: b
                  in: path
                  description: Right operand
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CalcMultiplyResponseBody'
                        required:
                            - result
            schemes:
                - http
    /openapi.json:
        get:
            tags:
                - calc
            summary: Download ./gen/http/openapi.json
            operationId: calc#/openapi.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
    /square_root:
        post:
            tags:
                - calc
            summary: SquareRoot calc
            operationId: calc#SquareRoot
            parameters:
                - name: object
                  in: body
                  required: true
                  schema:
                    type: object
                    properties:
                        a:
                            type: integer
                            description: number
                            example: 1187293393747494935
                            format: int64
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CalcSquareRootResponseBody'
                        required:
                            - result
            schemes:
                - http
    /subtraction/{a}/{b}:
        get:
            tags:
                - calc
            summary: subtraction calc
            operationId: calc#subtraction
            parameters:
                - name: a
                  in: path
                  description: Left operand
                  required: true
                  type: integer
                - name: b
                  in: path
                  description: Right operand
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CalcSubtractionResponseBody'
                        required:
                            - result
            schemes:
                - http
definitions:
    CalcAdditionResponseBody:
        title: CalcAdditionResponseBody
        type: object
        properties:
            result:
                type: integer
                example: 4285228716133201605
                format: int64
        example:
            result: 1232634558296710584
        required:
            - result
    CalcDivisionResponseBody:
        title: CalcDivisionResponseBody
        type: object
        properties:
            result:
                type: number
                example: 0.3215700892919487
                format: double
        example:
            result: 0.7075259881493255
        required:
            - result
    CalcMultiplyResponseBody:
        title: CalcMultiplyResponseBody
        type: object
        properties:
            result:
                type: integer
                example: 7617773024289378824
                format: int64
        example:
            result: 8614350958614154271
        required:
            - result
    CalcSquareRootResponseBody:
        title: CalcSquareRootResponseBody
        type: object
        properties:
            result:
                type: number
                example: 0.38201647214324563
                format: double
        example:
            result: 0.5949752326348026
        required:
            - result
    CalcSubtractionResponseBody:
        title: CalcSubtractionResponseBody
        type: object
        properties:
            result:
                type: integer
                example: 5635247262153220700
                format: int64
        example:
            result: 6595082841550375708
        required:
            - result